third_party/blink/web_tests/fast/forms/calendar-picker/week-picker-close-no-crash.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src=../resources/picker-common.js></script>

<input type=week id=week>

<script>
promise_test(() => {
  return openPicker(document.getElementById('week'))
    .then(() => {internals.pagePopupWindow.close()})
    .then(new Promise(resolve => setTimeout(resolve,0)))
    .then(() => {
      assert_equals(internals.pagePopupWindow, null, 'No window available');
    });
},'Accessing popup window controller after detach should not crash.');
</script>
</body>
</html>
